requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
cloudinary==1.28.0
cryptography==36.0.1
defusedxml==0.7.1
Deprecated==1.2.13
dj-database-url==0.5.0
Django==3.2.10
django-cloudinary-storage==0.3.0
django-cors-headers==3.10.1
django-environ==0.8.1
django-heroku==0.3.1
django-oauth-toolkit==1.6.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-social-oauth2==1.1.2
gunicorn==20.1.0
idna==3.3
jwcrypto==1.0
oauthlib==3.1.1
Pillow==8.4.0
psycopg2==2.9.2
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.7
whitenoise==5.3.0
wrapt==1.13.3
